pyproject.toml
[project]
name = "neuropaths"
version = "0.1.0"
description = "Learning transition paths in turbulent domains with neural operators (Fourier Neural Operators for committor functions of 2D advection-diffusion SDEs)."
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
    { name = "Tsz Fung Yu", email = "martin@tripudium.tech" },
]
keywords = [
    "neural-operator",
    "fourier-neural-operator",
    "committor-function",
    "transition-path-theory",
    "advection-diffusion",
    "turbulence",
]
classifiers = [
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Runtime deps: keep versions loose for now and pin via uv.lock.
# `autograd` is used specifically by the finite-difference solvers for
# coordinate-transform derivatives (NOT torch.autograd); keep it.
dependencies = [
    "torch>=2.3",
    "numpy>=1.26",
    "scipy>=1.13",
    "pandas>=2.2",
    "autograd>=1.7",
    "scikit-learn>=1.5",
    "matplotlib>=3.9",
    "pyyaml>=6.0",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.10",
]

[project.scripts]
# These entry points replace the bare `main()` calls in the legacy scripts.
# All three take `--config path/to/config.yaml` and write artefacts to the
# directory specified in the config (no hardcoded paths).
neuropaths-generate = "neuropaths.cli.generate:main"
neuropaths-train    = "neuropaths.cli.train:main"
neuropaths-evaluate = "neuropaths.cli.evaluate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/neuropaths"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
